+++ bindings/matrix-sdk-ffi/src/session_verification.rs
@@ -246,7 +246,10 @@ impl SessionVerificationController {
                     }
                     break;
                 }
-                SasState::Started { .. } | SasState::Accepted { .. } | SasState::Confirmed => (),
+                SasState::Created { .. }
+                | SasState::Started { .. }
+                | SasState::Accepted { .. }
+                | SasState::Confirmed => (),
             }
         }
     }